Move some remaining generic endpoints
